+++ NEWS.md
@@ -1,4 +1,9 @@
 
+# archeoViz 1.2.2
+Released: 2023-07-24
+
+* The 'Tables' tab is renammed 'Statistics'
+* Add export links to third-party applications (the Seriograph and Archeofrag applications) in the 'Statistics' tab.
 
 # archeoViz 1.2.1
 Released: 2023-06-28